Backend/MyntraAPI/requirements.txt
asgiref 
certifi==2020.12.5
dj-database-url==0.5.0
Django 
django-heroku==0.3.1
djangorestframework
gunicorn 
psycopg2 
PyJWT 
pytz==2020.1
sqlparse 
whitenoise 
requests
